+++ src/Nethermind/Nethermind.Core.Test/Json/DoubleConverterTests.cs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using System.Text.Json;
+using FluentAssertions;
 using Nethermind.Serialization.Json;
 using NUnit.Framework;
 
@@ -24,7 +25,7 @@ public class DoubleConverterTests
     public void Write_PreservesFullIeee754Precision(double value, string expected)
     {
         string json = JsonSerializer.Serialize(value, Options);
-        Assert.That(json, Is.EqualTo(expected));
+        json.Should().Be(expected, "double serialization must preserve full IEEE 754 round-trip precision");
     }
 
     [TestCase(0.678584082336891)]
@@ -35,6 +36,6 @@ public class DoubleConverterTests
     {
         string json = JsonSerializer.Serialize(value, Options);
         double deserialized = JsonSerializer.Deserialize<double>(json, Options);
-        Assert.That(deserialized, Is.EqualTo(value));
+        deserialized.Should().Be(value, "deserialization must recover the original double value exactly");
     }
 }
