+++ src/Nethermind/Nethermind.JsonRpc/Modules/Eth/FeeHistory/IFeeHistoryOracle.cs
@@ -7,6 +7,6 @@ namespace Nethermind.JsonRpc.Modules.Eth.FeeHistory
 {
     public interface IFeeHistoryOracle
     {
-        ResultWrapper<FeeHistoryResults> GetFeeHistory(int blockCount, BlockParameter newestBlock, double[]? rewardPercentiles);
+        ResultWrapper<FeeHistoryResults> GetFeeHistory(int blockCount, BlockParameter newestBlock, double[] rewardPercentiles);
     }
 }
